# backend/public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Notes API Documentation",
    "description": "RESTful API для управления заметками",
    "contact": {
      "email": "developer@example.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development Server"
    }
  ],
  "tags": [
    {
      "name": "Notes",
      "description": "Операции с заметками"
    }
  ],
  "paths": {
    "/api/notes": {
      "get": {
        "tags": ["Notes"],
        "summary": "Получить список всех заметок",
        "description": "Возвращает все заметки, отсортированные по дате создания",
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Note"
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Notes retrieved successfully"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Notes"],
        "summary": "Создать новую заметку",
        "description": "Создает новую заметку с валидацией данных",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Заметка успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Note"
                    },
                    "message": {
                      "type": "string",
                      "example": "Note created successfully"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/{id}": {
      "get": {
        "tags": ["Notes"],
        "summary": "Получить заметку по ID",
        "description": "Возвращает одну заметку по её идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID заметки",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Заметка найдена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Note"
                    },
                    "message": {
                      "type": "string",
                      "example": "Note retrieved successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Заметка не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Notes"],
        "summary": "Обновить заметку",
        "description": "Обновляет существующую заметку. Можно обновить отдельные поля.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID заметки",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Заметка обновлена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Note"
                    },
                    "message": {
                      "type": "string",
                      "example": "Note updated successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Заметка не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Notes"],
        "summary": "Удалить заметку",
        "description": "Удаляет заметку по её идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID заметки",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Заметка удалена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Note deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Заметка не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Note": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Моя заметка"
          },
          "content": {
            "type": "string",
            "example": "Содержимое заметки"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-12T17:56:21.000000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-12T17:56:21.000000Z"
          }
        }
      },
      "NoteInput": {
        "type": "object",
        "required": ["title", "content"],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "example": "Новая заметка"
          },
          "content": {
            "type": "string",
            "example": "Текст заметки"
          }
        }
      },
      "NoteUpdateInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "example": "Обновленный заголовок"
          },
          "content": {
            "type": "string",
            "example": "Обновленное содержимое"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "object",
            "example": {
              "title": ["The title field is required."],
              "content": ["The content field is required."]
            }
          }
        }
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Note not found"
          }
        }
      }
    }
  }
}